+++ kernel/include/arch/i386/ps2.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdint.h>
+#include <stddef.h>
 #include <arch/i386/idt.h>
 #include <util/option.h>
 
@@ -27,10 +28,10 @@ namespace ps2 {
 
     void init();
 
-    const Device& get_device(int id);
+    const Device& get_device(size_t id);
 
     /// 1 or 2.
-    int get_device_count();
+    size_t get_device_count();
 
     Option<const Device&> find_device_with_type(uint16_t type);
